+++ generator.go
@@ -1,3 +1,3 @@
-//go:generate go run ./codegen/codegen.go
+//go:generate go run ./pkg/codegen/codegen.go
 
 package main
