+++ pkg/rke/k8s_rke_system_images.go
