+++ pkg/rke/k8s_windows_service_options_test.go
