+++ pkg/rke/templates/aci-v5.2.3.5.go
