+++ pkg/rke/templates/aci-v5.2.3.go
