+++ pkg/rke/templates/calico.go
