+++ pkg/rke/templates/calico_v3.19.go
