+++ pkg/rke/templates/calico_v3.21.go
