+++ pkg/rke/templates/calico_v3.22-rancher2.go
