+++ pkg/rke/templates/calico_v3.22.go
