+++ pkg/rke/templates/calico_v3.24.go
