+++ pkg/rke/templates/calico_v3.25.go
