+++ pkg/rke/templates/canal.go
