+++ pkg/rke/templates/canal_v3.19.go
