+++ pkg/rke/templates/canal_v3.21.go
