+++ pkg/rke/templates/canal_v3.22-rancher2.go
