+++ pkg/rke/templates/canal_v3.22.go
