+++ pkg/rke/templates/canal_v3.24.go
