+++ pkg/rke/templates/canal_v3.25.go
