+++ pkg/rke/templates/coredns.go
