+++ pkg/rke/templates/coredns_v1.8.3-rancher2.go
