+++ pkg/rke/templates/flannel_v0.14.0-rancher2.go
