+++ pkg/rke/templates/kubedns.go
