+++ pkg/rke/templates/metrics.go
