+++ pkg/rke/templates/metrics_v0.5.0.go
