+++ pkg/rke/templates/metrics_v0.5.2.go
