+++ pkg/rke/templates/metrics_v0.6.1.go
