+++ pkg/rke/templates/nginx-v0.48.1.go
