+++ pkg/rke/templates/nginx-v0.49.3.go
