+++ pkg/rke/templates/nginx-v1.1.0-rancher3.go
