+++ pkg/rke/templates/nginx-v1.1.0.go
