+++ pkg/rke/templates/nginx-v1.2.0-rancher1.go
