+++ pkg/rke/templates/nginx-v1.5.1-rancher2.go
