+++ pkg/rke/templates/nginx-v1.7.0-rancher1.go
