+++ pkg/rke/templates/nodelocal.go
