+++ pkg/rke/templates/templates.go
