+++ pkg/rke/templates/weave.go
